multi-gene searches are working
