+++ app/controllers/api/v1/expression_violins_controller.rb
@@ -1,6 +1,6 @@
 module Api
   module V1
-    class ExpressionRendersController < ApiBaseController
+    class ExpressionViolinsController < ApiBaseController
       include Concerns::Authenticator
       include Concerns::StudyAware
       include Swagger::Blocks
@@ -18,7 +18,8 @@ module Api
         cluster = ApiParamUtils.get_cluster_group(params, @study)
         selected_annotation = ApiParamUtils.get_selected_annotation(params, @study, cluster)
         subsample = params[:subsample].blank? ? nil : params[:subsample].to_i
-        gene = @study.genes.by_name_or_id(params[:genes], @study.expression_matrix_files.map(&:id))
+        gene = @study.genes.by_name_or_id(params[:gene], @study.expression_matrix_files.map(&:id))
+
         render_data = ExpressionRenderingService.get_global_expression_render_data(
           @study, subsample, gene, cluster, selected_annotation, params[:boxpoints], current_user
         )
