+++ app/javascript/components/search/genes/GeneResultsPanel.js
@@ -5,7 +5,7 @@ import { faDna, faExclamationCircle } from '@fortawesome/free-solid-svg-icons'
 import { GeneSearchContext } from 'providers/GeneSearchProvider'
 import { StudyResults } from 'components/search/results/StudyResultsContainer'
 import { PagingControl } from 'components/search/results/PagingControl'
-import StudyGenes from './StudyGenes'
+import StudyGeneExpressions from './StudyGeneExpressions'
 
 /**
  * Component for Results displayed on the homepage
@@ -34,7 +34,7 @@ export default function GeneResultsPanel(props) {
         <StudyResults
           results={studyResults}
           changePage={pageNum => {searchContext.updateSearch({ genePage: pageNum })}}
-          StudyComponent={ StudyGenes }
+          StudyComponent={ StudyGeneExpressions }
         />
       </>
   } else {
