+++ app/javascript/styles/_resultsPanel.scss
@@ -48,6 +48,10 @@
   .loading-panel {
     color: #777
   }
+
+  .morpheus-nav {
+    display: none;
+  }
 }
 
 .result-row {
